scaffolding frontend application
- stub out the wasm ffi
- working example of running sqlsync in a web worker and communicating with the main thread :)
- make typescript happy
- working OPFS example code

+++ lib/sqlsync/src/journal/journal.rs
@@ -9,7 +9,7 @@ use crate::Serializable;
 
 use super::{Scannable, Syncable};
 
-pub type JournalId = i64;
+pub type JournalId = i32;
 
 #[derive(Error, Debug)]
 pub enum JournalError {
